Trigger Window restack with every focus events.
With the border overlapping feature added to the system, we need to make sure that we restack
windows with every focus update - Otherwise, some unfocused borders may hide the focus ones. This
comment ensures that whenever we call the focus method, we restack to ensure the focus window is on top of the other windows.